empresas-automatizadas-ia/exercicio-07-busca-semantica/requirements.txt
python-dotenv>=1.0.0,<2
fastapi>=0.115.0,<1
uvicorn[standard]>=0.32.0,<1
pydantic>=2.10.0,<3
numpy>=1.26.0,<3
torch>=2.2.0,<3
transformers>=4.40.0,<5
faiss-cpu>=1.8.0,<2
